Full funcs for admin

+++ app/Http/Controllers/API/CategoryController.php
@@ -2,9 +2,10 @@
 
 namespace App\Http\Controllers\API;
 
-use App\Http\Controllers\Controller;
 use App\Models\Category;
 use Illuminate\Http\Request;
+use App\Http\Controllers\Controller;
+use Illuminate\Support\Facades\Auth;
 
 class CategoryController extends Controller
 {
@@ -23,14 +24,20 @@ class CategoryController extends Controller
         }
     }
 
+
     public function store(Request $request)
     {
+        // Check if the user has the required role
+        if (!Auth::user()->roleUser || !Auth::user()->roleUser->role_id = 1 ){
+            return response()->json(['success' => false, 'message' => 'Unauthorized'], 401);
+        }
+    
         try {
             $validatedData = $request->validate([
                 'name' => 'required',
                 'slug' => 'required|unique:categories',
             ]);
-
+    
             $category = Category::create($validatedData);
             return response()->json(['success' => true, 'data' => $category], 201);
         } catch (\Exception $e) {
@@ -50,21 +57,31 @@ class CategoryController extends Controller
 
     public function update(Request $request, Category $category)
     {
+        // Check if the user has the required role
+        if (!Auth::user()->roleUser || !Auth::user()->roleUser->role_id = 1) {
+            return response()->json(['success' => false, 'message' => 'Unauthorized'], 401);
+        }
+    
         try {
             $validatedData = $request->validate([
                 'name' => 'required',
                 'slug' => 'required|unique:categories,slug,' . $category->id,
             ]);
-
+    
             $category->update($validatedData);
             return response()->json(['success' => true, 'data' => $category]);
         } catch (\Exception $e) {
             return response()->json(['success' => false, 'message' => $e->getMessage()]);
         }
     }
-
+    
     public function destroy(Category $category)
     {
+        // Check if the user has the required role
+        if (!Auth::user()->roleUser || !in_array(Auth::user()->roleUser->role_id, [2, 3])) {
+            return response()->json(['success' => false, 'message' => 'Unauthorized'], 401);
+        }
+    
         try {
             $category->delete();
             return response()->json(['success' => true], 204);
